Don't magicly call set_$key() from __set()
Calling a method leads to unexpected behaviour. Take the case of a
method set_password() in the object. Inside this method, you can
never call $this->password, as the call will simply loop back to
yourself.

It might have seemed like a good idea at some point, but it isn't
really.