get_details(): enforce numeric ids
This probably rarely happened anyway, but MySQL for example allows
querying for id (which is an int) by supplying a string, the string
would then be cast to an int and the matching record returned.

This means that fetching id "1a" would work just as well as fetching
id 1. This is probably unexpected, so this now throws an exception.